+++ Changelog.md
@@ -4,11 +4,25 @@
 
 ### Features and Improvements
 
-- None
+- Allow editing flow run names from the flow run page - [#448](https://github.com/PrefectHQ/ui/pull/448)
+- Allow editing task run names from the task run page - [#449](https://github.com/PrefectHQ/ui/pull/449)
 
 ### Bugfixes
 
 - None
+
+## 2020-11-13
+
+### Features and Improvements
+
+- Add feature teaser to the Artifacts tabs on the task run and flow run pages - [#435](https://github.com/PrefectHQ/ui/pull/435)
+
+## 2020-11-12
+
+### Features and Improvements
+
+- Remove the Gantt Chart (replaced by the Timeline) and `vue-apexcharts` - [#278](https://github.com/PrefectHQ/ui/pull/278)
+- Add the Flow Run timeline to the Flow Run Overview Page - [#278](https://github.com/PrefectHQ/ui/pull/278)
 
 ## 2020-11-11